[ADD] useSearchField custom hook && apply to SearchField Component
